+++ .sampo/changesets/issue-379-inspector-runtime-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/block-runtime: patch
+---
+
+Split inspector runtime view-model and rendering helpers into focused modules while preserving the public inspector export surface.
